requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.5
json-repair==0.58.6
streamlit==1.43.2
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
httpx==0.28.1
pytest==8.4.2
